Adds 'kw search'
Adds a new feature 'kw search' that recursively searches either a path
or a git log for a regex expression. If no arguments are given when
searching for a path, then it greps current directory. For 'kw search
log', if no arguments are given, then it prints whole git log. Also
added tests for this feature.

Signed-off-by: Giuliano Belinassi <giuliano.belinassi@usp.br>
Signed-off-by: Renato Lui Geh <renatogeh@gmail.com>